# mp-app/openapi.yaml
components:
  schemas:
...
    AutomationConfigIn:
      additionalProperties: false
      properties:
        enabled:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Enabled
        forbid_phi:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Forbid Phi
        secret:
          anyOf:
          - maxLength: 128
            minLength: 8
            type: string
          - type: 'null'
          title: Secret
        url:
          anyOf:
          - maxLength: 256
            type: string
          - type: 'null'
          title: Url
      title: AutomationConfigIn
      type: object
...
    BackupRequest:
      additionalProperties: false
      properties:
        password:
          maxLength: 200
          minLength: 8
          title: Password
          type: string
      required:
      - password
      title: BackupRequest
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    ManualCheckin:
      additionalProperties: false
      properties:
        member_id:
          title: Member Id
          type: integer
        method:
          default: manual
          title: Method
          type: string
      required:
      - member_id
      title: ManualCheckin
      type: object
...
    MemberPinLogin:
      additionalProperties: false
      properties:
        membership_code:
          maxLength: 64
          minLength: 1
          title: Membership Code
          type: string
        pin:
          maxLength: 32
          minLength: 1
          title: Pin
          type: string
      required:
      - membership_code
      - pin
      title: MemberPinLogin
      type: object
...
    PinLogin:
      additionalProperties: false
      properties:
        pin:
          maxLength: 32
          minLength: 1
          title: Pin
          type: string
        username:
          maxLength: 64
          minLength: 1
          title: Username
          type: string
      required:
      - username
      - pin
      title: PinLogin
      type: object
    QrCheckin:
      additionalProperties: false
      properties:
        method:
          default: qr
          title: Method
          type: string
        payload:
          additionalProperties: true
          title: Payload
          type: object
      required:
      - payload
      title: QrCheckin
      type: object
    RestoreRequest:
      additionalProperties: false
      properties:
        blob_b64:
          title: Blob B64
          type: string
        password:
          maxLength: 200
          minLength: 8
          title: Password
          type: string
      required:
      - password
      - blob_b64
      title: RestoreRequest
      type: object
...
    TokenResponse:
      properties:
        expires_in:
          title: Expires In
          type: integer
        gym_id:
          title: Gym Id
          type: integer
        role:
          title: Role
          type: string
        token:
          title: Token
          type: string
      required:
      - token
      - role
      - gym_id
      - expires_in
      title: TokenResponse
      type: object
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
...
info:
  description: Local-first Gym OS core API (offline-capable).
  title: Muscle Paradise Core
  version: 0.20.0
  x-generated-by: app.export_openapi
openapi: 3.1.0
paths:
  /api/v1/admin/backup:
    post:
      operationId: make_backup_api_v1_admin_backup_post
      parameters:
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BackupRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Make Backup Api V1 Admin Backup Post
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create a password-encrypted backup (OWNER)
      tags:
      - admin
  /api/v1/admin/backup/restore:
    post:
      operationId: restore_api_v1_admin_backup_restore_post
      parameters:
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RestoreRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Restore Api V1 Admin Backup Restore Post
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Restore from an encrypted backup, verifying row counts (OWNER)
      tags:
      - admin
...
  /api/v1/attendance/check-in:
    post:
      operationId: check_in_api_v1_attendance_check_in_post
      parameters:
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/QrCheckin'
              - $ref: '#/components/schemas/ManualCheckin'
              title: Body
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Check In Api V1 Attendance Check In Post
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Check in via signed QR or manual entry
      tags:
      - attendance
  /api/v1/attendance/check-out/{member_id}:
    post:
      operationId: check_out_api_v1_attendance_check_out__member_id__post
      parameters:
      - in: path
        name: member_id
        required: true
        schema:
          title: Member Id
          type: integer
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Check Out Api V1 Attendance Check Out  Member Id  Post
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Close the member's open visit
      tags:
      - attendance
  /api/v1/attendance/today:
    get:
      operationId: today_count_api_v1_attendance_today_get
      parameters:
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Today Count Api V1 Attendance Today Get
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Today's check-in count (ops KPI)
      tags:
      - attendance
  /api/v1/auth/me:
    get:
      description: Echo the verified identity — the shell renders role-appropriate UI.
      operationId: me_api_v1_auth_me_get
      parameters:
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Me Api V1 Auth Me Get
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Current principal
      tags:
      - auth
  /api/v1/auth/member-pin:
    post:
      description: 'Exchange a membership code + PIN for a MEMBER-scoped session token.


        Identical 401 for unknown code, unset PIN, and wrong PIN (no enumeration).'
      operationId: login_with_member_pin_api_v1_auth_member_pin_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberPinLogin'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Member PIN login
      tags:
      - auth
  /api/v1/auth/pin:
    post:
      description: "Exchange a staff username + PIN for a machine-local session token.\n\nRaises:\n  \
        \  HTTPException 401: unknown user, inactive user, or wrong PIN. The\n    message is deliberately\
        \ identical for all three (no user enumeration)."
      operationId: login_with_pin_api_v1_auth_pin_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PinLogin'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Staff PIN login
      tags:
      - auth
  /api/v1/automation/config:
    get:
      operationId: get_config_api_v1_automation_config_get
      parameters:
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Get Config Api V1 Automation Config Get
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Current bridge configuration (owner)
      tags:
      - automation
    post:
      operationId: update_config_api_v1_automation_config_post
      parameters:
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AutomationConfigIn'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Update Config Api V1 Automation Config Post
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Enable/disable the n8n bridge (owner)
      tags:
      - automation
...
  /api/v1/reports/expiring:
    get:
      operationId: expiring_api_v1_reports_expiring_get
      parameters:
      - in: query
        name: days
        required: false
        schema:
          default: 7
          maximum: 365
          minimum: 0
          title: Days
          type: integer
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  additionalProperties: true
                  type: object
                title: Response Expiring Api V1 Reports Expiring Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Members expiring within N days
      tags:
      - automation
  /api/v1/reports/inactive-members:
    get:
      operationId: inactive_api_v1_reports_inactive_members_get
      parameters:
      - in: query
        name: days
        required: false
        schema:
          default: 7
          maximum: 365
          minimum: 1
          title: Days
          type: integer
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  additionalProperties: true
                  type: object
                title: Response Inactive Api V1 Reports Inactive Members Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Members not seen in N days
      tags:
      - automation
...